Include Random Team Generator - Leverages @santacrab2's new team generator in ALM - Allows users to define a specific type team - Sends user a .zip with the legalized team of 6 to DM - Cleans up user command and bot responses to keep channels cleaned.

+++ SysBot.Pokemon/Helpers/AutoLegalityWrapper.cs
@@ -160,7 +160,7 @@ public static class AutoLegalityWrapper
         throw new ArgumentException("Type does not have a recognized trainer fetch.", typeof(T).Name);
     }
 
-    public static ITrainerInfo GetTrainerInfo(int gen) => TrainerSettings.GetSavedTrainerData((byte)gen);
+    public static ITrainerInfo GetTrainerInfo(byte gen) => TrainerSettings.GetSavedTrainerData(gen);
 
     public static PKM GetLegal(this ITrainerInfo sav, IBattleTemplate set, out string res)
     {
